+++ checker/src/main/java/org/checkerframework/checker/nullness/NullnessChecker.java
@@ -4,6 +4,7 @@ import org.checkerframework.checker.initialization.InitializationChecker;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.common.basetype.BaseTypeChecker;
 import org.checkerframework.common.basetype.BaseTypeVisitor;
+import org.checkerframework.framework.qual.StubFiles;
 import org.checkerframework.framework.source.SupportedLintOptions;
 
 import java.util.LinkedHashSet;
@@ -34,7 +35,12 @@ import javax.annotation.processing.SupportedOptions;
     NullnessChecker.LINT_TRUSTARRAYLENZERO,
     NullnessChecker.LINT_PERMITCLEARPROPERTY,
 })
-@SupportedOptions({"assumeKeyFor", "jspecifyNullMarkedAlias"})
+@SupportedOptions({
+    "assumeKeyFor",
+    "jspecifyNullMarkedAlias",
+    "conservativeArgumentNullnessAfterInvocation"
+})
+@StubFiles({"junit-assertions.astub"})
 public class NullnessChecker extends InitializationChecker {
 
     /** Should we be strict about initialization of {@link MonotonicNonNull} variables? */