+++ dataflow/src/main/java/org/checkerframework/dataflow/analysis/AnalysisResult.java
@@ -8,6 +8,7 @@ import org.checkerframework.checker.initialization.qual.UnknownInitialization;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.dataflow.cfg.block.Block;
 import org.checkerframework.dataflow.cfg.node.Node;
+import org.checkerframework.dataflow.util.UnmodifiableIdentityHashMap;
 import org.checkerframework.javacutil.BugInCF;
 import org.checkerframework.javacutil.TreeUtils;
 import org.plumelib.util.UniqueId;
@@ -32,8 +33,14 @@ import javax.lang.model.element.Element;
  */
 public class AnalysisResult<V extends AbstractValue<V>, S extends Store<S>> implements UniqueId {
 
+    /**
+     * For efficiency, certain maps stored in the result are only copied lazily, when they need to
+     * be mutated. This flag tracks if the copying has occurred.
+     */
+    private boolean mapsCopied = false;
+
     /** Abstract values of nodes. */
-    protected final IdentityHashMap<Node, V> nodeValues;
+    protected IdentityHashMap<Node, V> nodeValues;
 
     /**
      * Map from AST {@link Tree}s to sets of {@link Node}s.
@@ -41,13 +48,13 @@ public class AnalysisResult<V extends AbstractValue<V>, S extends Store<S>> impl
      * <p>Some of those Nodes might not be keys in {@link #nodeValues}. One reason is that the Node
      * is unreachable in the control flow graph, so dataflow never gave it a value.
      */
-    protected final IdentityHashMap<Tree, Set<Node>> treeLookup;
+    protected IdentityHashMap<Tree, Set<Node>> treeLookup;
 
     /**
      * Map from postfix increment or decrement trees that are AST {@link UnaryTree}s to the
      * synthetic tree that is {@code v + 1} or {@code v - 1}.
      */
-    protected final IdentityHashMap<UnaryTree, BinaryTree> postfixLookup;
+    protected IdentityHashMap<UnaryTree, BinaryTree> postfixLookup;
 
     /** Map from (effectively final) local variable elements to their abstract value. */
     protected final HashMap<Element, V> finalLocalValues;
@@ -91,9 +98,9 @@ public class AnalysisResult<V extends AbstractValue<V>, S extends Store<S>> impl
             IdentityHashMap<UnaryTree, BinaryTree> postfixLookup,
             HashMap<Element, V> finalLocalValues,
             Map<TransferInput<V, S>, IdentityHashMap<Node, TransferResult<V, S>>> analysisCaches) {
-        this.nodeValues = new IdentityHashMap<>(nodeValues);
-        this.treeLookup = new IdentityHashMap<>(treeLookup);
-        this.postfixLookup = new IdentityHashMap<>(postfixLookup);
+        this.nodeValues = UnmodifiableIdentityHashMap.wrap(nodeValues);
+        this.treeLookup = UnmodifiableIdentityHashMap.wrap(treeLookup);
+        this.postfixLookup = UnmodifiableIdentityHashMap.wrap(postfixLookup);
         // TODO: why are stores and finalLocalValues captured?
         this.stores = stores;
         this.finalLocalValues = finalLocalValues;
@@ -146,11 +153,22 @@ public class AnalysisResult<V extends AbstractValue<V>, S extends Store<S>> impl
      * @param other an analysis result to combine with this
      */
     public void combine(AnalysisResult<V, S> other) {
+        copyMapsIfNeeded();
         nodeValues.putAll(other.nodeValues);
         mergeTreeLookup(treeLookup, other.treeLookup);
         postfixLookup.putAll(other.postfixLookup);
         stores.putAll(other.stores);
         finalLocalValues.putAll(other.finalLocalValues);
+    }
+
+    /** Make copies of certain internal IdentityHashMaps, if they have not been copied already. */
+    private void copyMapsIfNeeded() {
+        if (!mapsCopied) {
+            nodeValues = new IdentityHashMap<>(nodeValues);
+            treeLookup = new IdentityHashMap<>(treeLookup);
+            postfixLookup = new IdentityHashMap<>(postfixLookup);
+            mapsCopied = true;
+        }
     }
 
     /**
@@ -406,12 +424,19 @@ public class AnalysisResult<V extends AbstractValue<V>, S extends Store<S>> impl
      *     running the analysis
      */
     protected @Nullable S runAnalysisFor(Node node, Analysis.BeforeOrAfter preOrPost) {
+        // block is null if node is a formal parameter of a method, or is a field access thereof
         Block block = node.getBlock();
-        assert block != null : "@AssumeAssertion(nullness): invariant";
+        assert block != null : "@AssumeAssertion(nullness): null block for node " + node;
         TransferInput<V, S> transferInput = stores.get(block);
         if (transferInput == null) {
             return null;
         }
+        // Calling Analysis.runAnalysisFor() may mutate the internal nodeValues map inside an
+        // AbstractAnalysis object, and by default the AnalysisResult constructor just wraps this
+        // map without copying it.  So here the AnalysisResult maps must be copied, to preserve
+        // them.
+        // TODO: Wouldn't it be safer to do at the beginning of the called method?
+        copyMapsIfNeeded();
         return runAnalysisFor(node, preOrPost, transferInput, nodeValues, analysisCaches);
     }