+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/visualize/CFGVisualizeLauncher.java
@@ -204,7 +204,7 @@ public final class CFGVisualizeLauncher {
 
         CFGVisualizer<V, S, T> viz = new DOTCFGVisualizer<>();
         viz.init(args);
-        Map<String, Object> res = viz.visualize(cfg, cfg.getEntryBlock(), analysis);
+        Map<String, Object> res = viz.visualizeWithAction(cfg, cfg.getEntryBlock(), analysis);
         viz.shutdown();
 
         if (pdf && res != null) {
@@ -230,17 +230,20 @@ public final class CFGVisualizeLauncher {
         Options.instance(context).put("compilePolicy", "ATTR_ONLY");
         JavaCompiler javac = new JavaCompiler(context);
 
-        JavacFileManager fileManager = (JavacFileManager) context.get(JavaFileManager.class);
-
-        JavaFileObject l =
-                fileManager.getJavaFileObjectsFromStrings(List.of(file)).iterator().next();
+        JavaFileObject l;
+        try (JavacFileManager fileManager = (JavacFileManager) context.get(JavaFileManager.class)) {
+            l = fileManager.getJavaFileObjectsFromStrings(List.of(file)).iterator().next();
+        } catch (IOException e) {
+            throw new Error(e);
+        }
 
         PrintStream err = System.err;
         try {
-            // redirect syserr to nothing (and prevent the compiler from issuing
-            // warnings about our exception.
+            // Redirect syserr to nothing (and prevent the compiler from issuing
+            // warnings about our exception).
             System.setErr(
                     new PrintStream(
+                            // In JDK 11+, this can be just "OutputStream.nullOutputStream()".
                             new OutputStream() {
                                 @Override
                                 public void write(int b) throws IOException {}