+++ framework/src/main/java/org/checkerframework/framework/ajava/InsertAjavaAnnotations.java
@@ -82,26 +82,30 @@ public class InsertAjavaAnnotations {
         }
 
         DiagnosticCollector<JavaFileObject> diagnostics = new DiagnosticCollector<JavaFileObject>();
-        JavaFileManager fileManager = compiler.getStandardFileManager(diagnostics, null, null);
-        if (fileManager == null) {
-            System.err.println("Could not get file manager");
-            System.exit(1);
-        }
+        try (JavaFileManager fileManager =
+                compiler.getStandardFileManager(diagnostics, null, null)) {
+            if (fileManager == null) {
+                System.err.println("Could not get file manager");
+                System.exit(1);
+            }
 
-        CompilationTask cTask =
-                compiler.getTask(
-                        null,
-                        fileManager,
-                        diagnostics,
-                        Collections.emptyList(),
-                        null,
-                        Collections.emptyList());
-        if (!(cTask instanceof JavacTask)) {
-            System.err.println("Could not get a valid JavacTask: " + cTask.getClass());
-            System.exit(1);
-        }
+            CompilationTask cTask =
+                    compiler.getTask(
+                            null,
+                            fileManager,
+                            diagnostics,
+                            Collections.emptyList(),
+                            null,
+                            Collections.emptyList());
+            if (!(cTask instanceof JavacTask)) {
+                System.err.println("Could not get a valid JavacTask: " + cTask.getClass());
+                System.exit(1);
+            }
 
-        return ((JavacTask) cTask).getElements();
+            return ((JavacTask) cTask).getElements();
+        } catch (IOException e) {
+            throw new Error(e);
+        }
     }
 
     /** Represents some text to be inserted at a file and its location. */
@@ -531,10 +535,11 @@ public class InsertAjavaAnnotations {
             File javaFile = new File(javaFilePath);
             String fileContents = FilesPlume.readFile(javaFile);
             String lineSeparator = FilesPlume.inferLineSeparator(annotationFilePath);
-            FileInputStream annotationInputStream = new FileInputStream(annotationFilePath);
-            String result = insertAnnotations(annotationInputStream, fileContents, lineSeparator);
-            annotationInputStream.close();
-            FilesPlume.writeFile(javaFile, result);
+            try (FileInputStream annotationInputStream = new FileInputStream(annotationFilePath)) {
+                String result =
+                        insertAnnotations(annotationInputStream, fileContents, lineSeparator);
+                FilesPlume.writeFile(javaFile, result);
+            }
         } catch (IOException e) {
             System.err.println(
                     "Failed to insert annotations from file "