+++ framework/src/main/java/org/checkerframework/framework/stub/JavaStubifier.java
@@ -191,7 +191,7 @@ public class JavaStubifier {
         @Override
         public MethodDeclaration visit(MethodDeclaration md, Void arg) {
             super.visit(md, arg);
-            // Non-private methods could be overriden with larger visibility.
+            // Non-private methods could be overridden with larger visibility.
             // So it is only safe to remove private methods, which can't be overridden.
             if (!removeIfPrivate(md)) {
                 md.removeBody();
@@ -236,7 +236,7 @@ public class JavaStubifier {
                 return false;
             }
             AccessSpecifier as = node.getAccessSpecifier();
-            if (as == AccessSpecifier.PRIVATE || as == AccessSpecifier.PACKAGE_PRIVATE) {
+            if (as == AccessSpecifier.PRIVATE || as == AccessSpecifier.NONE) {
                 ((Node) node).remove();
                 return true;
             }